+++ src/Checker.fs
@@ -2,7 +2,7 @@ namespace Kafka
 
 type Checker = {
     WaitForResourceDefault: int<second>
-    MaxRetries: int
+    MaxRetries: int<attempt>
     CheckCluster: Confluent.Kafka.Handle -> bool
     CheckTopic: StreamName -> Confluent.Kafka.Handle -> bool
 }
@@ -27,7 +27,7 @@ module Checker =
     let defaultChecker =
         {
             WaitForResourceDefault = 1<second>
-            MaxRetries = 10
+            MaxRetries = 10<attempt>
             CheckCluster = checkCluster
             CheckTopic = checkTopic
         }